requirements.txt
customtkinter
Pillow
python-dotenv
